+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Redux Framework Changelog
 
+## 3.3.9.17
+* Added     SASS compiler for admin CSS.
+* Added     SASS to CSS for grunt compiler.
+
+## 3.3.9.16
+* Added     SASS compiler for fields.
+
+## 3.3.9.16
+* Updated:  select2 3.5.2
+
 ## 3.3.9.15
 * Fixed:    #1856 - Illegal offset string warnings in framework.php.  Thanks @CGlingener.
 * Fixed:    #1874 - Required operator 'contains' and 'not_contains' not working with select set to multi.